% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfa_heatmap.R
\name{gfa_heatmap}
\alias{gfa_heatmap}
\title{GFA Heatmap.}
\usage{
gfa_heatmap(robW, block.names, varIdx.by.block, conf.level,
  heatmap.rep = FALSE, factor.order = NULL)
}
\arguments{
\item{robW}{No description.}

\item{block.names}{No description.}

\item{varIdx.by.block}{No description.}

\item{conf.level}{No description.}

\item{heatmap.rep}{No description.}

\item{factor.order}{No description.}
}
\description{
A function to produce a heatmap for robust factor loadings.
}
