% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOF.R
\name{GOF}
\alias{GOF}
\title{A function to compute goodness-of-fit on W matrix elements}
\usage{
GOF(est, true)
}
\arguments{
\item{est}{No description.}

\item{true}{No description.}
}
\description{
A function to compute goodness-of-fit on W matrix elements. It requires matching factors between estimated and true W matriices
}
