% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOF.rep.R
\name{GOF.rep}
\alias{GOF.rep}
\title{A function to compute goodness-of-fit on W matrix elements}
\usage{
GOF.rep(gfaList, parm, trueMtx)
}
\arguments{
\item{gfaList}{No description.}

\item{parm}{No description.}

\item{trueMtx}{No description.}
}
\description{
A function to compute goodness-of-fit on W matrix elements. It requires matching factors between estimated and true W matriices
}
