% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_dopar.R
\name{match_dopar}
\alias{match_dopar}
\title{A function to carry out matching in parallel.}
\usage{
match_dopar(comps, corGrids, matchGrids)
}
\arguments{
\item{comps}{A list of GFA replicates with posterior medians.}

\item{corGrids}{A list of cor thresholds to test. Cor threshold defines how close two components are required to be, in terms of correlation, in order to match them.}

\item{matchGrids}{A list of match thresholds to test. Match threshold describes the proportion of sampling chains that need to contain a component in order to include it in the robust components.}
}
\description{
A function to carry out matching in parallel.
}
