% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchComponents.R
\name{matchComponents}
\alias{matchComponents}
\title{Internal function for matching components.}
\usage{
matchComponents(comps, maxK, N, D, corThr, matchThr)
}
\arguments{
\item{comps}{No description.}

\item{maxK}{No description.}

\item{N}{No description.}

\item{D}{No description.}

\item{corThr}{No description.}

\item{matchThr}{No description.}
}
\description{
Revise robustComponents() to allow 0 matched factors.
}
