% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchFactors.R
\name{matchFactors}
\alias{matchFactors}
\title{A function to match factors.}
\usage{
matchFactors(comps, maxK, corThr, matchThr)
}
\arguments{
\item{comps}{No description.}

\item{maxK}{No description.}

\item{corThr}{No description.}

\item{matchThr}{No description.}
}
\description{
A function to match factors.
}
