% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE.Grids.R
\name{MSE.Grids}
\alias{MSE.Grids}
\title{A function to compute MSE of the GFA given robust components thresholds.}
\usage{
MSE.Grids(Ymtx, maxK, comps, corGrids, matchGrids)
}
\arguments{
\item{Ymtx}{The set of observed data as an NxD matrix.}

\item{maxK}{The maximal K among the GFA replicates.}

\item{comps}{A list of GFA replicates with posterior medians.}

\item{corGrids}{is a vector of cor thresholds to be tested. Cor threshold defines how close two components are required to be, in terms of correlation, in order to match them.}

\item{matchGrids}{is a vector of match thresholds to be tested. Match threshold describes the proportion of sampling chains that need to contain a component in order to include it in the robust components.}
}
\description{
A function to compute MSE b/w observed and reconstructed data matrices for each replicate for
each pair of (corThr, matchThr).
}
