% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizeK.R
\name{optimizeK}
\alias{optimizeK}
\title{A function to identify optimal (corThr, matchThr).}
\usage{
optimizeK(K.grids, mse.array)
}
\arguments{
\item{K.grids}{No description.}

\item{mse.array}{No description.}
}
\description{
A function to identify optimal (corThr, matchThr).
}
