% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmXW_by_factor.R
\name{pmXW_by_factor}
\alias{pmXW_by_factor}
\title{Compute factor-specific posterior mean of reconstructed data (X*t(W)).}
\usage{
pmXW_by_factor(gfa.obj)
}
\arguments{
\item{gfa.obj}{No description.}
}
\description{
Compute factor-specific posterior mean of reconstructed data (X*t(W)).
}
