% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psSummary.R
\name{psSummary}
\alias{psSummary}
\title{Functions to extract robust factors.}
\usage{
psSummary(gfa.res, credible.lv)
}
\arguments{
\item{gfa.res}{No description.}

\item{credible.lv}{No description.}
}
\description{
Functions to extract robust factors
(0) Using the GFA::gfa() resulting object to compute
(1) posterior medians
(2) credible interval limits
(3) factor-wise reconstratced data (i.e. X * W)
}
