% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob.var.exp.R
\name{rob.var.exp}
\alias{rob.var.exp}
\title{A function to compute \% variance explained}
\usage{
rob.var.exp(models, indices, block.names, varIdx.by.block,
  use.unmatched = F, by.block = T)
}
\arguments{
\item{models}{No description.}

\item{block.names}{No description.}

\item{varIdx.by.block}{No description.}

\item{use.unmatched}{No description.}

\item{by.block}{No description.}
}
\description{
A function to compute \% variance explained
}
