% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_wx.R
\name{rob_wx}
\alias{rob_wx}
\title{A function to compute robust loadings and robust scores.}
\usage{
rob_wx(models, indices, block.labs, var.labs = NULL)
}
\arguments{
\item{models}{No description.}

\item{indices}{No description.}

\item{block.labs}{No description.}

\item{var.labs}{No description.}
}
\description{
A function to compute robust loadings and robust scores.
}
