% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w.plot.R
\name{w.plot}
\alias{w.plot}
\title{Create a heatmap.}
\usage{
w.plot(w, D, K, gr1, conf.level, replicate)
}
\arguments{
\item{w}{No description.}

\item{D}{No description.}

\item{K}{No description.}

\item{gr1}{No description.}

\item{conf.level}{No description.}

\item{replicate}{No description.}
}
\description{
A base function to create a heat map.
}
