% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w.signs.R
\name{w.signs}
\alias{w.signs}
\title{Flip signs of factor loadings to ensure factors in the same direction across replicates.}
\usage{
w.signs(models, rob, use.unmatched = F)
}
\arguments{
\item{models}{No description.}

\item{rob}{No description.}

\item{use.unmatched}{No description.}
}
\description{
Flip signs of factor loadings to ensure factors in the same direction across replicates.
}
