% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/W_heatmap.R
\name{W_heatmap}
\alias{W_heatmap}
\title{For true W matrix parameters.}
\usage{
W_heatmap(W_DxK, varIdx.by.block, block.names)
}
\arguments{
\item{W_DxK}{No description.}

\item{varIdx.by.block}{No description.}

\item{block.names}{No description.}
}
\description{
For true W matrix parameters.
}
