% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.simu.R
\name{data.simu}
\alias{data.simu}
\title{For data simulation.}
\usage{
data.simu(N, W_DxK, varIdx.by.block, sd.noise)
}
\arguments{
\item{N}{sample size}

\item{W_DxK}{a list of matrices with K x D_{i} dimension}

\item{varIdx.by.block}{No description.}

\item{sd.noise}{SD of random Gaussian noise}
}
\description{
For data simulation.
}
