% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchFactors.R
\name{matchFactors}
\alias{matchFactors}
\title{A function to match factors.}
\usage{
matchFactors(comps, maxK, corThr, matchThr)
}
\arguments{
\item{comps}{A list of GFA replicates with posterior medians.}

\item{maxK}{The maximal K among the GFA replicates.}

\item{corThr}{Cor threshold defines how close two components are required to be, in terms of correlation, in order to match them.}

\item{matchThr}{Match threshold describes the proportion of sampling chains that need to contain a component in order to include it in the robust components.}
}
\description{
A function to match factors.
}
