% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE.Grids.R
\name{MSE.Grids}
\alias{MSE.Grids}
\title{A function to compute MSE.}
\usage{
MSE.Grids(Ymtx, maxK, comps, corGrids, matchGrids)
}
\arguments{
\item{Ymtx}{No description.}

\item{maxK}{No description.}

\item{comps}{No description.}

\item{corGrids}{No description.}

\item{matchGrids}{No description.}
}
\description{
A function to compute MSE b/w observed and reconstructed data matrices for each replicate for
each pair of (corThr, matchThr).
}
